jbNDT/requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
appdirs==1.4.4
arrow==1.2.3
asciitree==0.3.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
bidsschematools==0.6.0
blessed==1.20.0
botocore==1.27.59
cached-property==1.5.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
cryptography==42.0.5
cycler==0.11.0
dandi==0.55.0
dandischema==0.8.4
debugpy==1.7.0
decorator==5.1.1
dnspython==2.3.0
docker-pycreds==0.4.0
email-validator==2.0.0.post2
entrypoints==0.4
etelemetry==0.3.1
fasteners==0.19
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fscacher==0.4.0
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.42
h5py==3.8.0
hdmf==3.6.1
humanize==4.6.0
idna==3.6
importlib-metadata==4.2.0
importlib-resources==5.12.0
interleave==0.2.1
IProgress==0.4
ipykernel==6.16.2
ipython==7.34.0
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.19.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
keyring==23.9.3
keyrings.alt==4.2.0
kiwisolver==1.4.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
multidict==6.0.5
natsort==8.4.0
nest-asyncio==1.6.0
-e git+https://github.com/neurallatents/nlb_tools.git@42bb49f81585f7cfc849322d93127476f1c8af58#egg=nlb_tools
numcodecs==0.10.2
numpy==1.20.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nwbinspector==0.4.33
packaging==23.2
pandas==1.3.4
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.14
Pygments==2.17.2
pynwb==2.3.3
pyout==0.7.3
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3fs==2023.1.0
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==3.3.3
semantic-version==2.10.0
sentry-sdk==1.41.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tenacity==8.2.3
threadpoolctl==3.1.0
torch==1.13.1
torchsummary==1.5.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.18
wandb==0.16.4
wcwidth==0.2.13
webcolors==1.13
wrapt==1.16.0
yarl==1.9.4
zarr==2.12.0
zarr-checksum==0.4.0
zipp==3.15.0
